refactor: factorize AuthFilters to AbstractResourceTest
